cli: fix compilation error when CLI is not supported
The ccapi_streaming_cli.h header file is unconditionally using
connector_streaming_cli_* types which are defined by cc_ansic only when
CONNECTOR_STREAMING_CLI_SERVICE is defined.

To avoid the compilation errors when said symbol is not defined, use
conditional compilation in the code that directly or indirectly relies
on these types.

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>