CC: impliment oneshot command
Implimented the oneshot command for the cc. A callback is triggered
during the request stage that stores the command as is in a temporary
file. During the request stage the contents of the file are directly
given to the cli which sends its output to another file. Files are used
to prevent any weirdness with special characters.

These changes to how sessions are managed solve the error found in
DAL-4412.

cc_ansic changes have be synced.

DAL-4273, DAL-4412