CLI execute: remove extra character from payload
Message payload length was incremented one to many times resulting in a
garbage character being added to the payload.

Additionally, For messages that require multiple transmissions the status was
reinitialised to STREAMING_CLI_EXEC_STATUS_SUCCESS. Now we store it as a
session variable so it is preserved for multiple messages.

DAL-4273, DAL-4637, DAL-4631