allow to return status in ccapi response callback
Many of the builtin device_request functions expect the device to return a
stream name or otherwise return an SCI error with a non-zero status code.
cc_ansic supports that through the
connector_request_id_data_service_send_response callback, but that is only
called if a callback returns an error. Without this patch cc_api did not allow
the implementation layer callback to return an error and thus the status was
always 0 unless an internal error occurred.

DEL-7739

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>