dp: new function to remove older data point of each stream in a collection
In case the connector is running on a device that is not registered in any
account, data point collections grow and consume memory. Having this method
it is possible to remove old data points and only keep in memory a predefined
maximum of recent data points, deleting the older ones.

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>